+++ examples/Auth.php
@@ -1,0 +1,12 @@
+<?php
+
+use \PlugRoute\Middleware\PlugRouteMiddleware;
+use \PlugRoute\Http\Request;
+
+class Auth implements PlugRouteMiddleware
+{
+	public function handle(Request $request): Request
+	{
+		return $request;
+	}
+}