+++ examples/OtherMiddleware.php
@@ -1,0 +1,13 @@
+<?php
+
+use \PlugRoute\Middleware\PlugRouteMiddleware;
+use \PlugRoute\Http\Request;
+
+class OtherMiddleware implements PlugRouteMiddleware
+{
+	public function handle(Request $request): Request
+	{
+		$request->add('middleware', 'settting value');
+		return $request;
+	}
+}